Python code generation for MultiLayer extracted to separate class SampleToPython
